Move element components into components
