+++ src/components/image/index.ts
