+++ src/components/index.ts
@@ -1,11 +1,15 @@
 export { Badge } from './badge'
+export { Button } from './button'
 export { Card } from './card'
 export { Code } from './code'
 export { Header } from './header'
 export { Highlight } from './highlight'
-export { Markdown } from './markdown'
 export { Icon } from './icon'
+export { Image } from './image'
+export { Link } from './link'
+export { Markdown } from './markdown'
 export { Paper } from './paper'
+export { Preformat } from './preformat'
 export { Slideshow } from './slideshow'
 export { Social } from './social'
 export { Tooltip } from './tooltip'
