+++ src/components/markdown/markdown.scss
@@ -1,5 +1,5 @@
 @import '../../theme';
-@import '../../elements/preformat/preformat';
+@import '../preformat/preformat';
 
 :host {
   color: $text-color;
