--- src/elements/index.ts
@@ -1,4 +1,0 @@
-export { Button } from './button'
-export { Image } from './image'
-export { Link } from './link'
-export { Preformat } from './preformat'
