+++ src/index.ts
@@ -1,15 +1,11 @@
 // Dom Components
 import * as components from './components'
 
-// Element Components
-import * as elements from './elements'
-
 // Layout Components
 import * as layouts from './layouts'
 
 // Export the base elements objects
 export default {
   components,
-  elements,
   layouts,
 }
